core: update category

+++ app/Http/Controllers/CategoryController.php
@@ -8,8 +8,8 @@ use Illuminate\Contracts\View\View;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Nette\Utils\Random;
-use Psy\Readline\Hoa\Console;
+
+use function PHPUnit\Framework\returnSelf;
 
 class CategoryController extends Controller
 {
@@ -20,54 +20,45 @@ class CategoryController extends Controller
 
     public function store(Request $request): RedirectResponse {
         $input = $this->validate($request, [
-            'name' => 'required|string|max:255',
+            'name' => 'required|string|max:255'
         ]);
 
         try {
-            // dd($input);
             Category::create($input);
-
-            return redirect()->route('category.index')->with('success', 'Kategori berhasil ditambahkan');
+            return redirect()->route('category.index')->with('success', 'Data berhasil ditambahkan');
         } catch (\Throwable $th) {
-            return redirect()->back()->with('error', 'Kategori gagal ditambahkan');
+            return redirect()->back()->with('error', 'Data gagal ditambahkan');
         }
     }
 
-    public function update(Request $request, string $id): View | RedirectResponse {
+    public function update(Request $request, int $id): View | RedirectResponse {
         $mode = 'update';
 
         $category = Category::find($id);
-        if (!$category) return redirect()->back()->with('error', 'Category tidak ditemukan');
+        if (!$category) return redirect()->back()->with('error', 'Data tidak ditemukan');
 
         if ($request->getMethod() === 'GET') return view('pages.master.category.index', compact('mode', 'category'));
 
         $input = $this->validate($request, [
-            'name' => 'required|string|max:255',
+            'name' => 'required|string|max:255'
         ]);
 
         try {
-            $category->fill($input)->save();
-            return redirect()->route('category.index')->with('success', 'Kategori berhasil ditambahkan');
+            $category->update($input);
+            return redirect()->route('category.index')->with('success', 'Data berhasil diubah');
         } catch (\Throwable $th) {
-            return redirect()->back()->with('error', 'Kategori gagal ditambahkan');
+            return redirect()->back()->with('error', 'Data gagal diubah');
         }
     }
 
-    // public function destroy(Request $request): JsonResponse{
-    //     dd($request->all());
-    //     $category = Category::find($request->id);
-    //     if (!$category) return response()->json([ 'message' => 'Data tidak ditemukan' ], 404);
-
-    //     $category->delete();
-    //     return response()->json([ 'message' => 'Data berhasil dihapus' ]);
-    // }
-
-    public function destroy(Request $request): JsonResponse{
+    public function destroy(Request $request): JsonResponse {
         $category = Category::find($request->id);
-        if (!$category) return response()->json([ 'message' => 'Data tidak ditemukan' ], 404);
+        if (!$category) return response()->json(['message' => 'Data tidak ditemukan'], 404);
+
+        if (Product::where('category_id', $category->id)->exists()) return response()->json(['message' => 'Data tidak dapat dihapus karena masih digunakan'], 400);
 
         $category->delete();
-        return response()->json([ 'message' => 'Data berhasil dihapus' ]);
+        return response()->json(['message' => 'Data berhasil dihapus'], 200);
     }
 
     public function query(Request $request): JsonResponse {
@@ -77,23 +68,8 @@ class CategoryController extends Controller
         $order = $request->order?? 'desc';
         $orderBy = $request->orderBy?? 'created_at';
 
-        // $category = Category::leftJoin('products', 'products.category_id', '=', 'categories.id')
-        //     ->select('categories.id as category_id', 'categories.name', 'products.available_qty as qty', 'categories.created_at', 'categories.updated_at')
-        //     ->orderBy($orderBy, $order);
-
-        $category = Category::leftJoin('products', 'categories.id', '=', 'products.category_id')
-        ->selectRaw('categories.id as category_id, categories.name, 
-            IFNULL(SUM(products.available_qty), 0) as qty, 
-            categories.created_at, categories.updated_at')
-        ->groupBy('categories.id')
-        ->orderBy('categories.created_at', 'desc');
-
-        // $category = Category::leftJoin('products', 'products.category_id', '=', 'categories.id')
-        //     ->selectRaw('categories.id as category_id, categories.name, categories.created_at, categories.updated_at')
-        //     ->groupBy('categories.id')
-        //     ->orderBy('categories.created_at', 'desc');
-
-        // $data['category'] = $category->get();
+        $category = Category::select('id', 'name', 'created_at', 'updated_at')
+            ->orderBy($orderBy, $order);
 
         if ($limit && is_numeric($limit))   $category->limit($limit);
         if ($offset && is_numeric($offset)) $category->offset($offset);
